Re-work the way that we load file data for python modules.
It's now loaded "on demand" when the interpreter sees an import
statement, using an async xhr request.  This means that the necessary
files are written into the filesystem only once it's clear they'll
be needed.

It needs a lot of polish, but seems to work ok and means we don't have
to use synchronous XHR for the filesystem.