Adds remove as an option to module_bundler->purge
Adds --remove to purge command.  If the module is preloaded, it completely removes it (removes module content from index.json, and removes import tree entry).  If the module is not preloaded, it removes it from the import tree, but does not touch the underlying file inside the module cache (that file can then be removed via rm).  If the entry points to a directory, only the top level entry is removed, each sub-entry can be removed by a separate entry.  

Probably should support some form of wild-carding, but I'm not sure exactly what form that should be... (simple bash-like? automatically assume all sub modules should be removed? full-on re.match based matching?)